Add Endpoint for Publishing a Tale+Data
Adds an endpoint, `createPackage`, that takes a list of item ids, the tale, and a data repository.

Internally, it sorts the files into ones that exist remotely on DataONE and ones the exist on the local filesystem.

Files on the local FS have suitable metadata generated for them.

Adds supoprt for handling files on the girder filesystem.

Note that I didn't werite tests for the dataone_package file. There are a number of changes that the file may go through in upcoming commits.