Add Support for Uploading EML &  a Resource Map Describing DataONE Objects
This commit extends the endpoint to take a list of n DataONE linked files, create an EML document, and add the linked file pids to the resource map, which is then uploaded to a member node.

Note that most of these changes are uncommenting a few lines and changing variable names.