Dynamic Node Endpoints
I created an enum to hold the different node endpoints and put it in the utility file.

I added a parameter to the
   1. lookupData
   2. listFiles
endpoints. The parameter takes a member/cn endpoint that will be used for locating the file on the network and registering it in Whole Tale.

I was having issues finding files from dev.nceas with the stage-2 coordinating node, so I added a reference to the member node (which finds them without issue). I also added a new regex for the view page of dev.nceas. I could make it more general, combing the search for `search.dataone's` view page, by looking for `#view`, but I'm not sure if we want it so genreal.

Because the endpoint can now be specified by the user, any call to `dataone_register.query` had to be modified to take a url so you'll see a host of functions that now take `base_url`.

The goal is to let the user set the endpoint in the UI and have their input trickle in to the backend interactions. This should be generic enough to support the UI.

I also moved `utils.filter_items` to `dataone_upload.py`

Changed the import style to match the rest of the project (I was listing each file before)