Add File Describing Paths
This commit regards adding a file to the data package that describes where each file came from on the WT system. This is needed so that when we import the package back into Whole Tale, we can create a proper directory structure.

In utils.py:
  I added a function that removes keys from a dictionary.

In dataone_upload.py:
  I created a fiunction, `create_upload_file_paths` that is responsible for creating the paths file and uploading it to DataONE. When the EML is generated for thisfile, the size is needed so I propagated this change in `create_upload_eml`

In dataone_package.py
  Made changes to add a section of EML that describes the paths file.

I also changed a few model access levels to READ instead of ADMIN.