Add License Information to EML
We need to let the user pick a license for their package. This adds a parameter to the package endpoint. The licenses are keyed in constants.py

Most of the change can be read by first looking at the changes in constants.py, then going to repository.py and following the call stack. You'll see that the license id persists up to EML generation.

This corresponds to issue #128