Support Uploading DataONE Files
# Change 1: Moved Publishing Endpoint
I created a new endpoint, `publish`, with DataONE as `publish/dataone`. I would anticipate globus being `publish/globus`

# Change 2: I moved Static File Descriptions to Constants
The files that we inject into the packages (LICENSE, tale.yaml, etc) have their descriptions in `constants.py`.

# Change 3:Use `with` For TempFile in create_external_object_structure
Also added a check to see if the file downloaded

# Change 4: Transfer DataONE Objects to NCEAS
This is done in `transfer_prod_to_dev`. We loop over each item, download it into a temporary file, hash it, and then upload it ito DataONE. The importants bits about the uploaded object are stored in a dict, which are referenced in the EML creation code. These are in the `new_dataone_objects` that is passed around in `dataone_upload`.